+++ Kint-Sources-master/OpenNos.GameObject/Packets/ClientPackets/MviPacket.cs
@@ -17,10 +17,10 @@ namespace OpenNos.GameObject
         public short Slot { get; set; }
 
         [PacketIndex(2)]
-        public byte Amount { get; set; }
+        public short Amount { get; set; }
 
         [PacketIndex(3)]
-        public byte DestinationSlot { get; set; }
+        public short DestinationSlot { get; set; }
 
         #endregion
     }
